Added second map "Frozen Lakes"
Added a second map with a button on the stage select.

Has two paths, where spawned enemies alternate between the two paths. The path chosen first is randomized when the map is selected. (Note: the paths' length are approximately the same length as the first map's path).

Additionally, enemies now despawn and deal damage as soon as they reach the base.