Resize game to 1920x1080
Resized all rooms in the game to be 1920px by 1080px, per the rubric.

Adjusted maps to fit the new size, and moved the bar and its data to the right side instead of the bottom.

Due to the paths of the Frozen Lakes map becoming shorter, enemy speed is set to 85% for balancing.

I also left the old versions of the maps in the files in case we need them still.