Moved bar items to right
Moved all items previously in the bottom bar to the new bar placement on the right. Increased size of money and round values.

I also fixed an issue with the bar switcher that caused lots of lag over time. Along with this, I moved everything from obj_bar into obj_player and made a script to switch the bar info.