Added shell alien
Added a shell alien, which starts in a shell, and upon losing half health loses its shell.

With its shell, it resists damage from bullets by half. It is immune to the flamethrower, and a rocket explosion will destroy the shell earlier.

Without its shell, it goes twice as fast, but deals half the damage.

Also added a button for the rocket tower.