Small changes + crash fixes
Cheats that alter gameplay now check for obj_player so it doesn't crash when used on the title screen.

Set rounds that are incomplete to empty arrays, so it doesn't crash when trying to load the round.

Changed center of all enemy sprites to bottom center, and moved all paths down by 32. This makes most aliens move the same as before, but the king alien now looks better.

Also fixed alignment on the game's logo.