Setup for end screen
Added global variables that track if the game is won, the total money spent, and the total towers placed. These statistics (+ final health  remaining) will be displayed on the end screen.

Also set up viewports on both maps, viewport 0 being the default view and viewport 1 excluding the side bar, for the end screen.